Add builder for panic handler and doc

+++ xpanic/handlepanic.go
@@ -9,28 +9,39 @@ import (
 	"syscall"
 )
 
-// HandlePanic writes a message to the logger and causes a backtrace to be produced.
 // If crashOnError is set a coredump will be produced and kill program else it continues.
-func HandlePanic(Errorf func(format string, args ...interface{}), crashOnError bool) {
-	r := recover()
-	if r != nil {
-		Errorf("capture panic infos")
+const (
+	CrashOnErrorActivated = true
+	CrashOnErrorDeactivated = false
+)
 
-		Errorf(fmt.Sprintf("panic: %s", r))
-		Backtrace(Errorf)
+// BuildPanicHandler builds a panic handler and verifies a none nil logger got passed
+func BuildPanicHandler(errorf func(format string, args ...interface{}), crashOnError bool) func() {
+	if errorf == nil {
+		panic("errorf is not set")
+	}
+	// handlePanic writes a message to the logger and causes a backtrace to be produced.
+	return func() {
+		r := recover()
+		if r != nil {
+			errorf("capture panic infos")
 
-		if crashOnError {
-			signal.Reset(syscall.SIGABRT)
-			Errorf("finished capturing of panic infos")
-			syscall.Kill(0, syscall.SIGABRT)
-		} else {
-			Errorf("finished capturing of panic infos")
+			errorf(fmt.Sprintf("panic: %s", r))
+			Backtrace(errorf)
+
+			if crashOnError {
+				signal.Reset(syscall.SIGABRT)
+				errorf("finished capturing of panic infos")
+				syscall.Kill(0, syscall.SIGABRT)
+			} else {
+				errorf("finished capturing of panic infos")
+			}
 		}
 	}
 }
 
 // Backtrace writes a multi-line backtrace to the logger.
-func Backtrace(Errorf func(format string, args ...interface{})) {
+func Backtrace(errorf func(format string, args ...interface{})) {
 	profiles := pprof.Profiles()
 	buf := new(bytes.Buffer)
 
@@ -38,11 +49,11 @@ func Backtrace(Errorf func(format string, args ...interface{})) {
 		// https://golang.org/pkg/runtime/pprof/#Profile.WriteTo.
 		err := pprof.Lookup(p.Name()).WriteTo(buf, 2)
 		if err != nil {
-			Errorf("could not write profile: %v", err)
+			errorf("could not write profile: %v", err)
 		}
 	}
 
 	for _, line := range strings.Split(buf.String(), "\n") {
-		Errorf(line)
+		errorf(line)
 	}
 }
